+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version 1.3.7 - Community 1.0.0
+
+    * [SER-3502] - Send country when submitting the account transfer bulk upload
+    
 ## Version 1.3.6 - Community 1.0.0
 
     * [SER-3363] - Implement label translation in Client related UIs